+++ flink-end-to-end-tests/flink-end-to-end-tests-common/src/main/java/org/apache/flink/tests/util/cache/AbstractDownloadCache.java
@@ -18,10 +18,10 @@
 
 package org.apache.flink.tests.util.cache;
 
+import org.apache.flink.test.parameters.ParameterProperty;
 import org.apache.flink.tests.util.AutoClosableProcess;
 import org.apache.flink.tests.util.CommandLineWrapper;
 import org.apache.flink.tests.util.TestUtils;
-import org.apache.flink.tests.util.parameters.ParameterProperty;
 import org.apache.flink.util.TimeUtils;
 
 import org.apache.commons.io.FileUtils;
