+++ flink-end-to-end-tests/flink-end-to-end-tests-common/src/main/java/org/apache/flink/tests/util/flink/LocalStandaloneFlinkResourceFactory.java
@@ -18,8 +18,8 @@
 
 package org.apache.flink.tests.util.flink;
 
-import org.apache.flink.tests.util.parameters.ParameterProperty;
-import org.apache.flink.tests.util.util.FileUtils;
+import org.apache.flink.test.parameters.ParameterProperty;
+import org.apache.flink.test.util.FileUtils;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
